+++ packages/base-modal/src/Component.tsx
@@ -12,6 +12,7 @@ import React, {
     useMemo,
     useRef,
     useState,
+    WheelEvent,
 } from 'react';
 import FocusLock from 'react-focus-lock';
 import mergeRefs from 'react-merge-refs';
@@ -205,9 +206,19 @@ export type BaseModalProps = {
     componentRef?: MutableRefObject<HTMLDivElement | null>;
 
     /**
+     * Реф контентной области
+     */
+    contentElementRef?: MutableRefObject<HTMLDivElement | null>;
+
+    /**
      * Блокирует скролл когда модальное окно открыто. Работает только на iOS.
      */
     iOSLock?: boolean;
+
+    /**
+     * Хэндлер события прокрутки колесиком
+     */
+    onWheel?: (e: WheelEvent<HTMLElement>) => void;
 };
 
 export type BaseModalContext = {
@@ -274,8 +285,10 @@ export const BaseModal = forwardRef<HTMLDivElement, BaseModalProps>(
             dataTestId,
             zIndex = stackingOrder.MODAL,
             componentRef = null,
+            contentElementRef = null,
             usePortal = true,
             iOSLock = false,
+            onWheel,
         },
         ref,
     ) => {
@@ -590,6 +603,7 @@ export const BaseModal = forwardRef<HTMLDivElement, BaseModalProps>(
                                 onKeyDown={handleKeyDown}
                                 onMouseDown={handleBackdropMouseDown}
                                 onMouseUp={handleBackdropMouseUp}
+                                onWheel={onWheel}
                                 tabIndex={-1}
                                 data-test-id={dataTestId}
                                 style={{
@@ -625,7 +639,12 @@ export const BaseModal = forwardRef<HTMLDivElement, BaseModalProps>(
                                                 styles.content,
                                                 contentClassName,
                                                 contentProps?.className,
+                                                {
+                                                    [styles.hasFooter]: hasFooter,
+                                                    [styles.hasHeader]: hasHeader,
+                                                },
                                             )}
+                                            ref={contentElementRef}
                                         >
                                             {children}
                                         </div>