+++ packages/circular-progress-bar/src/shared/index.ts
@@ -1,0 +1,1 @@
+export { getCircularProgressBarTestIds } from '../utils/get-circular-progress-bar-test-ids';