+++ packages/custom-button/src/desktop/index.ts
@@ -1,0 +1,1 @@
+export { CustomButtonDesktop } from './Component.desktop';