+++ packages/universal-date-input/src/components/date-input/Component.tsx
@@ -8,6 +8,7 @@ import { Input } from '@alfalab/core-components-input';
 import { Popover } from '@alfalab/core-components-popover';
 import { getDataTestId } from '@alfalab/core-components-shared';
 import { CalendarMIcon } from '@alfalab/icons-glyph/CalendarMIcon';
+import { CalendarSIcon } from '@alfalab/icons-glyph/CalendarSIcon';
 
 import {
     DATE_FORMAT,
@@ -207,6 +208,8 @@ export const DateInput = forwardRef<HTMLInputElement, InnerDateInputProps>(
             return null;
         };
 
+        const CalendarIcon = restProps.size === 40 ? CalendarSIcon : CalendarMIcon;
+
         return (
             <div
                 {...wrapperHandlers}
@@ -236,9 +239,12 @@ export const DateInput = forwardRef<HTMLInputElement, InnerDateInputProps>(
                         <React.Fragment>
                             {rightAddons}
                             {picker && (
-                                <CalendarMIcon
+                                <CalendarIcon
                                     onClick={onPickerClick}
-                                    className={styles.calendarIcon}
+                                    className={cn(
+                                        styles.calendarIcon,
+                                        styles[`size-${restProps.size}`],
+                                    )}
                                     onMouseDown={preventDefault}
                                 />
                             )}